Address Copilot review round 7
- Clear stale wikidata_sitelinks in single-actor backfill mode (set NULL
  instead of only updating timestamp)
- Cast omdb_box_office_cents to float instead of numeric so pg returns a
  JS number rather than a string
- Apply same ::float cast in backfill-dof-popularity-direct for consistency

+++ server/scripts/backfill-dof-popularity-direct.ts
@@ -125,7 +125,7 @@ async function backfillMovies(
     const result = await pool.query(
       `
       SELECT tmdb_id, title, release_year, original_language,
-             omdb_box_office_cents::numeric as omdb_box_office_cents,
+             omdb_box_office_cents::float as omdb_box_office_cents,
              trakt_watchers, trakt_plays,
              omdb_imdb_votes, tmdb_popularity, production_countries,
              omdb_awards_wins, omdb_awards_nominations, aggregate_score
